backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
duckdb==0.10.2
pandas==2.2.2
groq==0.7.0
openai==1.30.0
upstash-redis==1.1.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
aiosqlite==0.20.0
sqlalchemy==2.0.30
python-multipart==0.0.9
